requirements.txt
fastapi
uvicorn
gradio
duckdb
google-generativeai
python-dotenv
requests
